+++ packages/python-packages/apiview-stub-generator/tests/init_files/__init__.extend_w_comments_docs.py
@@ -1,0 +1,25 @@
+# ---------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# ---------------------------------------------------------
+"""Contains core packages, modules, and classes for Azure Machine Learning SDKv2.
+
+Main area includes the following:
+* SDK objects/entities are data classes describing AzureML concepts and configurations.
+* Operations are classes contain logic to interact with backend services, usually auto generated operations call.
+* Autorest objects.
+* SDK v2 support the new YAML file format.
+"""
+__path__ = __import__("pkgutil").extend_path(__path__, __name__)
+
+# This is a multi-line docstring with single quotes
+'''
+Multi-line docstring w/ single quotes
+'''
+# This is a multi-line docstring with double quotes
+"""
+Multi-line docstring w/ double quotes
+"""
+# This is a single-line docstring with single quotes
+'''Single line docstring w/ single quotes'''
+# This is a single-line docstring with double quotes
+"""Single line docstring w/ single quotes"""