+++ src/java/apiview-java-processor/src/main/java/com/azure/tools/apiview/processor/analysers/JavaASTAnalyser.java
@@ -95,6 +95,7 @@ public class JavaASTAnalyser implements Analyser {
 
     private final Map<String,String> shortClassNameToFullyQualfiedNameMap = new HashMap<>();
 
+    private int JAVADOC_LINE_COUNT = 0;
 
 
     /************************************************************************************************
@@ -391,7 +392,7 @@ public class JavaASTAnalyser implements Analyser {
         // parent
         String gavStr = mavenPom.getParent().getGroupId() + ":" + mavenPom.getParent().getArtifactId() + ":"
                 + mavenPom.getParent().getVersion();
-        MiscUtils.tokeniseMavenKeyValue(mavenLine, "parent", gavStr);
+        MiscUtils.tokeniseMavenKeyValue(mavenLine, "parent", gavStr, true);
 
         // properties
         gavStr = mavenPom.getGroupId() + ":" + mavenPom.getArtifactId() + ":" + mavenPom.getVersion();
@@ -1463,7 +1464,8 @@ public class JavaASTAnalyser implements Analyser {
         }
 
         Stream.of(lines).forEach(line -> {
-            final ReviewLine reviewLine = parent.addChildLine().setRelatedToLine(targetLineId);
+            // we add a line id here so that each line of the javadoc can be commented on
+            final ReviewLine reviewLine = parent.addChildLine("JAVADOC_LINE_" + JAVADOC_LINE_COUNT++).setRelatedToLine(targetLineId);
 
             if (line.contains("&")) {
                 line = HtmlEscape.unescapeHtml(line);