+++ packages/python-packages/apiview-stub-generator/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.3.17 (Unreleased)
 Fixed issue with azure-core dependency conflict when running the stub generator on azure-core by moving necessary core files into an internal vendor folder.
+Fixed bug where package namespace was not being discovered correctly due to missing checks for init extension file format.
 
 ## Version 0.3.16 (2025-03-03)
 Fixed emty package name issue when running parser against pacakge source path instead of wheel. PKG_INFO is not available in this case.